[layout] Move Kani proofs from crate root
All moved proofs test code from the `layout` module, and so belong there
rather than in the crate root.

gherrit-pr-id: I691b42ce8c0c3c6e5990e7684fc66f8f5dd73d85
